Extract CatchLuredPokemonWorker from PokemonCatchWorker and improved worker order (#1627)
* extracted lure catch worker from pokemon catch

* removing information less logs
